create server node back-end set auth with firebase create token for sign in sign up create check jwt token
