+++ src/.eslintrc.json
@@ -1,0 +1,9 @@
+{
+    "parserOptions": {
+        "ecmaVersion": "latest"
+    },
+
+    "env": {
+        "es6": true
+    }
+}
