+++ src/environments/environment.prod.ts
@@ -9,4 +9,13 @@ export const environment = {
     messagingSenderId: "559886303180",
     appId: "1:559886303180:web:3b1125c27eba7ae7926769",
   },
+  firebaseConfig :{
+    apiKey: "AIzaSyD2wmPaAlLVy7iGY6L0gZt_i3MlC2vXb_A",
+    authDomain: "spotytest-e89c6.firebaseapp.com",
+    projectId: "spotytest-e89c6",
+    storageBucket: "spotytest-e89c6.appspot.com",
+    messagingSenderId: "823395277840",
+    appId: "1:823395277840:web:4988a37c1d18697de2fe95",
+    measurementId: "G-47B0D4P2EH"
+  },
 };
