+++ src/environments/environment.ts
@@ -5,6 +5,15 @@
 export const environment = {
   production: false,
   url_api: 'http://esiea-spotymike.eu-4.evennode.com/v1',
+  firebaseConfig:{
+    apiKey: "AIzaSyD2wmPaAlLVy7iGY6L0gZt_i3MlC2vXb_A",
+    authDomain: "spotytest-e89c6.firebaseapp.com",
+    projectId: "spotytest-e89c6",
+    storageBucket: "spotytest-e89c6.appspot.com",
+    messagingSenderId: "823395277840",
+    appId: "1:823395277840:web:4988a37c1d18697de2fe95",
+    measurementId: "G-47B0D4P2EH"
+  },
 };
 
 /*
