+++ src/server/serviceAccountKey.json
@@ -1,0 +1,13 @@
+{
+    "type": "service_account",
+    "project_id": "spotytest-e89c6",
+    "private_key_id": "7ab83e6cd8409cf306006165c58fd6520dd632fd",
+    "private_key": "-----BEGIN PRIVATE KEY-----\nMIIEvAIBADANBgkqhkiG9w0BAQEFAASCBKYwggSiAgEAAoIBAQCxvgNjvHCyP0rg\n4kgHfRAbZXx7SN/W328F+O0eXLFpfpoedWsRaIi1R71qsafn1Tr9GjhfE1+kocXk\nM9hASTNqnIYdtsyBnE7RdnXpaotmXCPKuKRQnLgzi9kZxXclZycrP0uc3ekmr1DS\n5WXVC33fr+TIbIcTrEmBDHUQhRCnHSiObXrPi/8/Cgsu4PQcF3FL4NyX/Z6C41V3\n5T9fAtgRJGZ2CInlt/3TLH/zRKZitaD009KWyS3Wn+E5jgL3tUf0e5jVxiOaJtDc\nFeVj8lkCQg4KcT1N0TsOYNaAckvlZGOPJu6RHogoKJo15Hd4LceM1lyKHiJKefME\n55oa3MZfAgMBAAECggEAJnsBR9v5uBVs1CAZQQnJ7Wx81UITN/Val3vrdhnSTQkc\nO8DnA7dSdZDJCOrac54d7cL59ld+6ukdK0kwJcbTS18EyDUVtG/PPpj7cSPgdzI2\ni4KLDGrJlFtQDrzKDT52Y+4JpDYE7puhD2+bqtKYl0+A5AVe/xrbNQWp6IAueUdA\n1lDixAl/ILemZc8cSa6rWXvql5PwWUmS4OmVhfh+G1OyDM7mGU39PfrcxSY/QJs+\nZZrpQcvvQdRa+ZXLsPB//rOadKNzM7mVoCw7EczBOryXOK0iA/iFUhDRxO9pQ4ZH\nqOsV2e6jQlfonmXKlOb1tVRplxnnZEizoRlOFDm0iQKBgQD4u/Hwmjx+YJHLsH45\nEO59UgMsNug70VgGvd4fFFf5i9ZXLk2XwIS/EX8oj3wQdSOE9qWHRNFi9R2R1aGB\n3fBjrP538FVn0wH1zDNg6SmZ3rrnjWa4fDzs/kk4F57TB+HrhlVkr718Ew8qQZKp\nToe+cWCMzA9lKZRkZeVeOLcN9wKBgQC27y9DNhc3SWUvChc85rbeMhUdcqy4fezY\nA7J6xnVs09AIeA8LgoG/mcFtcBahqJDyDiAGXa7qr4Bdcly5tKUXrfOz7Cm6oQhO\nQrR1fEs9kt+InkmcaGFWfL5NDMusjxqcfs4K3i+LGpvAgoZ9mSJZLQgc4lLcA6y7\noefhFSGQ2QKBgDa2oPGb+9KQy4FSwXEsGEnY0HnZWhxOYF6bRVDw2Kk76VokBU1s\nO8cLwjzZ0yCLo9SLrL2klap+4JJDu9MlTsueMhuP8jl0YOoAkQywPUZriBGHOSE7\nsusuxExiZb1l7nwwgHg1cCjqyFszMV+sto9dwS83c+t8yfeLsniXYSzhAoGARcv6\nv20q7Hg/d/Qkq6LnkN2Q+txfFEFPlZfLiFaiRPSIAkM0klKoCJ+0knna2LFFJGXJ\nSESEUZzGoGDbCivndaJ0QmjP/xoj3iBLVCwpVQPt7+Vc9qVIxkPwI9vG9j7fi2hC\nWxktEKkFVi02/ULpPRcWalbcAIASbsC2hUWXufkCgYAORDWAbpzk4HmYXQNUqsPt\nUvRJzuwJdSsiDvnup0pN41vBV3sTQ9QmhV36FjuU3XFifmT1/6/R12BwETmtf0bW\nvoolBB0rJT03hFXZK0LHN7yHypvE0rtoCt40E/9NQeBaHQFVPXbqt4VG/tsNdZ4k\n0LuxqPUOXcfTYcEU0lGqhg==\n-----END PRIVATE KEY-----\n",
+    "client_email": "firebase-adminsdk-h4idf@spotytest-e89c6.iam.gserviceaccount.com",
+    "client_id": "110574889532454970031",
+    "auth_uri": "https://accounts.google.com/o/oauth2/auth",
+    "token_uri": "https://oauth2.googleapis.com/token",
+    "auth_provider_x509_cert_url": "https://www.googleapis.com/oauth2/v1/certs",
+    "client_x509_cert_url": "https://www.googleapis.com/robot/v1/metadata/x509/firebase-adminsdk-h4idf%40spotytest-e89c6.iam.gserviceaccount.com",
+    "universe_domain": "googleapis.com"
+  }
